move FallbackManager to inner class; add callbacks
